refactor(publicodes-state): introduce NGCQuestionType and use switch statement instead of if-then-else
