+++ src/publicodes-state/hooks/useRule/useType.ts
@@ -1,8 +1,8 @@
 'use client'
 
 import { useMemo } from 'react'
-import getType from '../../helpers/getType'
-import { NGCEvaluatedNode, NGCRuleNode } from '../../types'
+import getType from '../helpers/getType'
+import { NGCEvaluatedNode, NGCQuestionType, NGCRuleNode } from '../types'
 
 type Props = {
   dottedName: string
@@ -13,9 +13,7 @@ type Props = {
 export default function useType({ dottedName, rule, evaluation }: Props) {
   // Model shenanigans
 
-  const type = useMemo<
-    'notQuestion' | 'mosaic' | 'choices' | 'boolean' | 'number' | undefined
-  >(() => {
+  const type = useMemo<NGCQuestionType | undefined>(() => {
     return getType({ dottedName, rule, evaluation })
   }, [dottedName, rule, evaluation])
